El fichero quijote tiene 65527 chars (bytes), lo maximo. Prueba que si pones uno mas ya no te deja. He modificado su "funcion" calcular_checksum haciendo que el primer argumento sea en vez de un uint16_t un uint32_t porque sino se desborda para checksums de segmentos longitudes mayores a 2^16 - 12. Amen.

+++ P4/src/practica4.h
@@ -82,7 +82,7 @@ uint8_t inicializarPilaEnviar();
 uint8_t registrarProtocolo(uint16_t protocolo, pf_notificacion handleModule, pf_notificacion* protocolos_registrados);
 
 /***************************Funciones auxiliares implementadas***********************************/
-uint8_t calcularChecksum(uint16_t longitud, uint8_t *datos, uint8_t *checksum);
+uint8_t calcularChecksum(uint32_t longitud, uint8_t *datos, uint8_t *checksum);
 uint8_t mostrarPaquete(uint8_t * paquete, uint32_t longitud);
 void handleSignal(int nsignal);
